[feat]: [PL-25585]: Use Events Framework for NG Ldap Group sync (#36582)
* [feat]: [PL-25585]: Update hashcode


* [feat]: [PL-25585]: Update with develop


* [feat]: [PL-25585]: Fixed getLdap settings from CG for existing settings.


* [feat]: [PL-25585]: Updated endpoints for getting ldap settings from CG


* [feat]: [PL-25585]: Fixed Test cases


* fix: [CDS-41642]: Fix NullPointerException in `aggregateStatus` method of GraphRenderer class (#36386)

* [CDS-41642]: Fix null pointer exception in aggregateStatus method

* [CDS-41642]: Fix codeformat

* Address suggestion

* [CDS-41642]: Log when stateExecutionInstance status with null value is being aggregated

* fix: [CI-4477]: remove prints for gitlab refresh token (#36528)

* [fix]: [CDS-39326]: fixing helmchart filter (#36305)

* fixing helmchart

* fix lint

* feat: [PL-27456]: Rename the target to be module and make run analysis per module (#36488)

* Rename the target to be module and make run analysis per module

* Cleaned up code after code review comments.

* feat: [CI-5283]: setting pod timeout for hosted/free users.  (#36463)

* Setting pod timeout to be 1 hour for hosted free users

* Unit test

* fix

* set active time for direct

* move ttl extraction to the helper

* remove unused import

* removed unused field

* Code owner for build-cleaner. (#36536)

* fix: [PL-27456]: Include empty deps list in BUILD files (#36455)

* fix: [CCM-8659]: Update Anomaly Details Links (#36486)

* fix: [CCM-8659]: update anomaly details links

* fix: [CCM-8659]: codeformat

* fix: [CCM-8659]: unused variables removed

* fix: [CCM-8336]: Recommendation doc update (#36411)

* fix: [CCM-8336]: recommendation doc update

* fix: [CCM-8336]: Recommendation doc update codeformat

* feat: [CI-4692]: Git clone step implementation (#34679)

* feat: [CI-4692]: Git Clone Step - Initial implementation

* feat: [CI-4692]: Git Clone Step - Resolve code review comments

* feat: [CI-4692]: Git Clone Step - PR check resolution

* feat: [CI-4692]: Git Clone Step - When no cloneDirectory provided, default to /harness/repoName. Do not allow /harness to be used as a git clone step clone directory.

* feat: [CI-4692]: Git Clone Step - Maintain drone env variables for all containers except git clone step containers

* feat: [CI-4692]: Git Clone Step - add exception for null build environment variables

* feat: [CI-4692]: Git Clone Step - clear codebase env variables in GitClone method of plugin setting utils for VM.

* feat: [CI-4692]: Git Clone Step - code review comment improvements

* feat: [CI-4692]: Git Clone Step - PR check PMD issue fixed

* Revert "[CDS-41724]: Adding traits for CD license usage in CG" (#36539)

This reverts commit ce53f1322ff5aba5a032fe482328ca132646f79f.

* [feat]: [PL-25585]: sync groups trigger through event

* [feat]: [PL-27534]: Ldap test authentication API in NG

(cherry picked from commit 8c7daa31bb0725c8feeb31a408fc619c7759700c)

* [feat]: [PL-27534]: Ldap test authentication API in NG

(cherry picked from commit 197735428c1c0f46ad0e6946f9da60aa38df268a)

* [feat]: [PL-25585]: Added delete check and ng group link publish event.


* [feat]: [PL-25585]: Added delete check and ng group link publish event.


* [feat]: [PL-25585]: Added delete check and ng group link publish event.


* [feat]: [PL-25585]: Remove sync on NG Group link


* [feat]: [PL-25585]: Fix tests


* [feat]: [PL-25585]: Fix style checks


* [feat]: [PL-25585]: Fix review comments


* [feat]: [PL-25585]: Removed ngLdapGroupSyncEventPublisher rom NG as its not required


* [feat]: [PL-25585]: Removed unnecessary comments


* [feat]: [PL-25585]: Removed unnecessary comments


* [feat]: [PL-25585]: add anotation for feature restriction check

+++ 122-ng-authentication-settings/src/main/java/io/harness/ng/authenticationsettings/impl/AuthenticationSettingsServiceImpl.java
@@ -281,14 +281,17 @@ public class AuthenticationSettingsServiceImpl implements AuthenticationSettings
   }
 
   @Override
-  public LDAPSettings createLdapSettings(String accountIdentifier, LDAPSettings ldapSettings) {
+  @FeatureRestrictionCheck(FeatureRestrictionName.LDAP_SUPPORT)
+  public LDAPSettings createLdapSettings(
+      @NotNull @AccountIdentifier String accountIdentifier, LDAPSettings ldapSettings) {
     log.info("NGLDAP: Create ldap settings call for accountId {}", accountIdentifier);
     return fromCGLdapSettings(getResponse(
         managerClient.createLdapSettings(accountIdentifier, toCGLdapSettings(ldapSettings, accountIdentifier))));
   }
 
   @Override
-  public LDAPSettings updateLdapSettings(String accountIdentifier, LDAPSettings ldapSettings) {
+  public LDAPSettings updateLdapSettings(
+      @NotNull @AccountIdentifier String accountIdentifier, LDAPSettings ldapSettings) {
     log.info("NGLDAP: Update ldap settings call for accountId {}, ldap name {}", accountIdentifier,
         ldapSettings.getDisplayName());
     return fromCGLdapSettings(getResponse(
@@ -296,13 +299,22 @@ public class AuthenticationSettingsServiceImpl implements AuthenticationSettings
   }
 
   @Override
-  public void deleteLdapSettings(String accountIdentifier) {
+  public void deleteLdapSettings(@NotNull @AccountIdentifier String accountIdentifier) {
     log.info("NGLDAP: Delete ldap settings call for accountId {}", accountIdentifier);
+    LdapSettings settings = getResponse(managerClient.getLdapSettings(accountIdentifier));
+    if (settings == null) {
+      throw new InvalidRequestException("No Ldap Settings found for this account: " + accountIdentifier);
+    }
+    if (isNotEmpty(userGroupService.getUserGroupsBySsoId(accountIdentifier, settings.getUuid()))) {
+      throw new InvalidRequestException(
+          "Deleting Ldap provider with linked user groups is not allowed. Unlink the user groups first");
+    }
     getResponse(managerClient.deleteLdapSettings(accountIdentifier));
   }
 
   @Override
-  public LdapResponse testLDAPLogin(@NotNull String accountIdentifier, String email, String password) {
+  public LdapResponse testLDAPLogin(
+      @NotNull @AccountIdentifier String accountIdentifier, String email, String password) {
     log.info("NGLDAP: Test ldap authentication in accountId {}", accountIdentifier);
     return getResponse(managerClient.testLdapAuthentication(
         accountIdentifier, createPartFromString(email), createPartFromString(password)));
